Prevent registered disposables throwing on cleanup
When using a component which extends `PropertyBinder` and utilized `new LucidEventTarget(this)`, the event target disposable was getting disposed of during the test on the destruction of the component, and the component was then trying to dispose of the event target **again** in this method. This fix prevents already disposed of registered disposables from attempting to dispose of twice.